# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Blog Post Api with crud operations",
    "version": "1.0.0",
    "title": "Blog Post Api",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "tags": [
    {
      "name": "Products API",
      "description": "Test API"
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
      "Users": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "public_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "admin": {
            "type": "string"
          }
        }
      },
      "Blog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "blog": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "comments": {
            "type": "string"
          }
        }
      },
      "Comments": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "post_id": {
            "type": "string"
          }
        }
      },
      "Id": {
        "type": "integer",
        "properties": {
          "name": {
            "type": "integer"
          }
        }
      }
    }
  },
  "paths": {
    "/user": {
      "post": {
        "tags": [
          "Add User to API"
        ],
        "summary": "Add user",
        "requestBody": {
          "description": "Post New User To Database",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/user"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/user"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Get All Users from API"
        ],
        "summary": "Get All User",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "Get user from API"
        ],
        "summary": "Get User",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "user id to update",
            "required": true,
            "type": "integer",
            "schema": {
              "$ref": "#/components/schemas/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Update User on API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id to update",
            "required": true,
            "type": "integer",
            "schema": {
              "$ref": "#/components/schemas/user"
            }
          }
        ],
        "summary": "Update USer",
        "requestBody": {
          "description": "Update Existing User On Database",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/user"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Delete user to API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "user id to delete",
            "required": true,
            "type": "integer",
            "schema": {
              "$ref": "#/components/schemas/user"
            }
          }
        ],
        "summary": "Delete User",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/user"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/user"
            }
          }
        }
      }
    }
  }

}